+++ hier_config/platforms/cisco_ios/view.py
@@ -173,7 +173,7 @@ class ConfigViewInterfaceCiscoIOS(  # noqa: PLR0904
 
     @property
     def port_number(self) -> int:
-        return int(self.name.split("/")[-1].split(".")[0])
+        return int(self.number.split("/")[-1].split(".")[0])
 
     @property
     def speed(self) -> tuple[int, ...] | None: